improve css, move some things around
