+++ packages/editor/src/components/properties/PropertyGroup.tsx
@@ -9,8 +9,8 @@ const StyledPropertyGroup = styled.div`
   flex-direction: column;
   flex-shrink: 0;
   width: 100%;
-  padding: 12px 0;
-  border-bottom: 1px solid var(--border);
+  padding-top: 4px;
+  border-top: 1px solid var(--border);
 `
 
 // PropertyGroupHeader used to provide styles for property group header
@@ -20,7 +20,7 @@ const PropertyGroupHeader = styled.div`
   align-items: left;
   font-weight: bold;
   color: var(--textColor);
-  padding: 0 8px 8px;
+  padding: 0 8px;
   :last-child {
     margin-left: auto;
   }
@@ -30,14 +30,13 @@ const PropertyGroupHeader = styled.div`
 const PropertyGroupDescription = styled.div`
   color: var(--textColor);
   white-space: pre-wrap;
-  padding: 0 8px 8px;
+  padding: 0 8px;
 `
 
 // component to contain content of property group
 const PropertyGroupContent = styled.div`
   display: flex;
   flex-direction: column;
-  height: 20px;
 `
 
 // component to contain content of property group
@@ -47,13 +46,13 @@ const PropertyCloseButton = styled.button`
   margin: auto;
   margin-right: 0px;
   font-size: 16px;
-  height: 20px;
+  height: 16px;
   background: none;
   border: none;
   color: var(--iconButtonColor);
 
   &:hover {
-    font-size: 20px;
+    font-size: 14px;
     color: var(--iconButtonHoverColor);
   }
